feat : Add 메모리 공간 부족으로 malloc 실패 시, 이전에 할당했던 공간을 free 하는 함수
