+++ BornToCode/push_swap/ver1.3/push_swap/ft_deque_util.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_deque_list.c                                    :+:      :+:    :+:   */
+/*   ft_deque_util.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: kimjaehwan <marvin@42.fr>                  +#+  +:+       +#+        */
+/*   By: jaekim <jaekim@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/05/13 15:43:58 by kimjaehwa         #+#    #+#             */
-/*   Updated: 2022/05/13 15:43:59 by kimjaehwa        ###   ########.fr       */
+/*   Created: 2022/05/13 18:52:59 by jaekim            #+#    #+#             */
+/*   Updated: 2022/05/13 18:53:01 by jaekim           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
