+++ BornToCode/push_swap/ver1.3/push_swap/ft_rotate.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_rotate.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: kimjaehwan <marvin@42.fr>                  +#+  +:+       +#+        */
+/*   By: jaekim <jaekim@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/05/13 15:41:26 by kimjaehwa         #+#    #+#             */
-/*   Updated: 2022/05/13 15:41:27 by kimjaehwa        ###   ########.fr       */
+/*   Created: 2022/05/13 18:52:41 by jaekim            #+#    #+#             */
+/*   Updated: 2022/05/13 18:52:43 by jaekim           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
