+++ BornToCode/push_swap/ver1.3/push_swap/ft_split.c
@@ -3,72 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_split.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: kimjaehwan <marvin@42.fr>                  +#+  +:+       +#+        */
+/*   By: jaekim <jaekim@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/05/13 15:38:55 by kimjaehwa         #+#    #+#             */
-/*   Updated: 2022/05/13 15:39:03 by kimjaehwa        ###   ########.fr       */
+/*   Created: 2022/05/13 18:53:12 by jaekim            #+#    #+#             */
+/*   Updated: 2022/05/13 18:53:13 by jaekim           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "push_swap.h"
 
-int	count_size(char **av)
-{
-	int	i;
-	int	j;
-	int	length;
-
-	i = 1;
-	length = 0;
-	while (0 != av[i])
-	{
-		j = 0;
-		while (0 != av[i][j])
-		{
-			if ((9 <= av[i][j] && 13 >= av[i][j]) || 32 == av[i][j])
-				j++;
-			else
-			{
-				while (0 != av[i][j]
-				 && (9 > av[i][j] || 13 < av[i][j]) && 32 != av[i][j])
-					j++;
-				length++;
-			}
-		}
-		i++;
-	}
-	return (length);
-}
-
-void	make_str(char *word, char *av, int position, int length)
-{
-	int		i;
-
-	i = 0;
-	while (i < length)
-	{
-		word[i] = av[position];
-		position++;
-		i++;
-	}
-	word[i] = '\0';
-}
-
-int	check_size(char **av, int i, int j)
-{
-	while ((9 <= av[i][j] && 13 >= av[i][j]) || 32 == av[i][j])
-	{
-		j++;
-		if (0 == av[i][j])
-			return (0);
-	}
-	while (('0' <= av[i][j] && '9' >= av[i][j])
-		|| '+' == av[i][j] || '-' == av[i][j])
-		j++;
-	return (j);
-}
-
-char	**ft_second_split(char **result, char **av)
+void	ft_second_split(char **result, char **av)
 {
 	int		i;
 	int		j;
@@ -83,16 +27,10 @@ char	**ft_second_split(char **result, char **av)
 		while (0 != av[i][j])
 		{
 			position = j;
-			if (0 != check_size(av, i, j))
-			{
-				j = check_size(av, i, j);
-				result[index] = (char *)malloc(sizeof(char) * (j - position + 1));
-			}
-			if (0 == result[index])
-			{
-				index--;
+			j = check_size(av, i, j);
+			creat_result(result, &index, &j, &position);
+			if (check_index(result, index))
 				break ;
-			}
 			make_str(result[index], av[i], position, j - position);
 			if (0 == av[i][j])
 				break ;
@@ -101,7 +39,6 @@ char	**ft_second_split(char **result, char **av)
 		index++;
 		i++;
 	}
-	return (result);
 }
 
 char	**ft_split(char **av)
