+++ BornToCode/push_swap/ver1.3/push_swap/ft_swap.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_swap.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: kimjaehwan <marvin@42.fr>                  +#+  +:+       +#+        */
+/*   By: jaekim <jaekim@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/05/13 15:23:41 by kimjaehwa         #+#    #+#             */
-/*   Updated: 2022/05/13 15:23:42 by kimjaehwa        ###   ########.fr       */
+/*   Created: 2022/05/13 18:51:26 by jaekim            #+#    #+#             */
+/*   Updated: 2022/05/13 18:51:28 by jaekim           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
