+++ BornToCode/push_swap/ver1.3/push_swap/push_swap.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   push_swap.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: kimjaehwan <marvin@42.fr>                  +#+  +:+       +#+        */
+/*   By: jaekim <jaekim@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/05/13 15:24:04 by kimjaehwa         #+#    #+#             */
-/*   Updated: 2022/05/13 15:59:23 by kimjaehwa        ###   ########.fr       */
+/*   Created: 2022/05/13 18:52:15 by jaekim            #+#    #+#             */
+/*   Updated: 2022/05/13 18:52:16 by jaekim           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,16 +32,27 @@ typedef struct s_deque
 }					t_deque;
 
 void	push_swap(t_deque *deque_a, t_deque *deque_b);
+void	a_to_b(t_deque *deque_a, t_deque *deque_b);
+void	b_to_a(t_deque *deque_a, t_deque *deque_b);
+int		check_sorting(t_deque *deque_a);
+void	sort_small_size(t_deque *deque_a, t_deque *deque_b, int size);
+
 int		ft_printf(const char *argu, ...);
 int		ft_atoi(const char *str);
 char	**ft_split(char **av);
+int		count_size(char **av);
+void	make_str(char *word, char *av, int position, int length);
+int		check_size(char **av, int i, int j);
+int		check_index(char **result, int index);
+void	creat_result(char **result, int *index, int *j, int *position);
 size_t	ft_strlen(const char *str);
+int		check_sorting(t_deque *deque_a);
 
 int		check_null(char **av);
 int		check_spaces(char **av);
 int		check_digit(char **av);
 int		check_sign(char **av);
-int		check_error(char **av);
+void	check_error(int ac, char **av);
 int		*check_error_and_split(int ac, char **av);
 
 void	ft_make_list(int *av, t_deque *deque_a, t_node **node_a);
@@ -49,32 +60,25 @@ void	ft_make_index(t_node *node_a);
 void	ft_make_list_index(int *av, t_deque *deque_a, t_node *node_a);
 
 int		ft_lstsize(t_node *lst);
+void	ft_lstclear(t_node **lst);
 void	ft_renew_lstsize(t_deque *deque_a, t_deque *deque_b);
 void	ft_lstadd_back(t_node **lst, t_node *new);
 void	ft_deq_init(t_deque **deque_a, t_deque **deque_b);
 void	ft_deqadd_tail(t_deque *deque, t_node *node_a);
 void	ft_deqadd_head(t_deque *deque, t_node *node_a);
 
-void	print_lst_deq(t_deque *deque);
-void	print_node(t_node *node);
-void	print_all(t_deque *deque_a, t_deque *deque_b);
-
 void	ft_swap_a(t_deque *deque_a);
 void	ft_swap_b(t_deque *deque_b);
 void	ft_swap_ab(t_deque *deque_a, t_deque *deque_b);
-
 void	ft_push_a(t_deque *deque_a, t_deque *deque_b);
 void	ft_push_b(t_deque *deque_a, t_deque *deque_b);
-
 void	ft_rotate_a(t_deque *deque_a);
 void	ft_rotate_b(t_deque *deque_b);
 void	ft_rotate_ab(t_deque *deque_a, t_deque *deque_b);
-
 void	ft_reverse_rotate_a(t_deque *deque_a);
 void	ft_reverse_rotate_b(t_deque *deque_b);
 void	ft_reverse_rotate_ab(t_deque *deque_a, t_deque *deque_b);
 
-void	a_to_b(t_deque *deque_a, t_deque *deque_b);
 t_node	*ft_lstlast(t_node *lst);
 t_node	*ft_lstnew(int value);
 t_deque	*ft_deqnew(t_deque *lst);
